scripts/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
vnstock>=3.4.0
